Boundary Points Visualization

+++ server.js
@@ -100,13 +100,37 @@ app.post("/api/visualize_homography", upload.single("file"), (request, response,
     return;
   }
 
+  // TODO: Check for boundary image OR base-map image. 
+  if (!fs.existsSync("./resources/visualize-homography/cricket_map.png")) {
+    error_obj = {
+      "message": "Please add the image of the standard cricket map for visualizing homography!",
+      "code": 404
+    }
+    response.status(404).send({error: error_obj});
+    return;
+  }
+
   // Save the incoming image points in the visualize-homography directory
-  image_points = JSON.parse(request.body.data)["image-points"];
-  fs.mkdir("./resources/visualize-homography", {recursive: true}, no_op);
+  dataObj = JSON.parse(request.body.data);
+  image_points = dataObj["image-points"];
+  // fs.mkdir("./resources/visualize-homography", {recursive: true}, no_op);
   fs.writeFileSync("./resources/visualize-homography/image_points.json", JSON.stringify({ "points": image_points }))
 
+  // TODO: Conditional construction of the exec command as a string. 
+  base_map_path = "./resources/visualize-homography/cricket_map.png";
+  if(!dataObj["is-boundary"]) {
+    if (fs.existsSync("./resources/visualize-homography/boundary_map.png"))
+      base_map_path = "./resources/visualize-homography/boundary_map.png";
+  }
+  
+  var options = {
+    "mode": "apply_homography",
+    "image_points_path": "./resources/visualize-homography/image_points.json",
+
+  }
+  let execCmd = "python3 ./resources/homography.py --mode apply_homography --image_points_path ./resources/visualize-homography/image_points.json --homography_path ./resources/compute-homography/homography_matrix.npy --frame_path ./resources/uploads/ground_frame.png --map_path " + base_map_path + " --output_dir ./resources/visualize-homography " + (dataObj["is-boundary"] === true ? "--is_boundary" : ""); 
   // Trigger the visualization script. 
-  shell.exec("python3 ./resources/homography.py --mode apply_homography --image_points_path ./resources/visualize-homography/image_points.json --homography_path ./resources/compute-homography/homography_matrix.npy --frame_path ./resources/uploads/ground_frame.png --map_path ./resources/visualize-homography/cricket_map.png --output_dir ./resources/visualize-homography", {async: false});
+  shell.exec(execCmd, {async: false});
 
   response_obj = {
     "message": "Execution completed!!",
